Extract all storages into separate classes and inject `DatabasePool` into them
