+++ UnstoppableWallet/UnstoppableWallet/Core/Managers/LogRecordManager.swift
@@ -2,9 +2,9 @@ import HsToolKit
 
 class LogRecordManager: ILogRecordManager {
     private let linesCountLimit: Int
-    private let storage: ILogRecordStorage
+    private let storage: LogRecordStorage
 
-    init(storage: ILogRecordStorage, linesCountLimit: Int = 1000) {
+    init(storage: LogRecordStorage, linesCountLimit: Int = 1000) {
         self.storage = storage
         self.linesCountLimit = linesCountLimit
     }
