+++ UnstoppableWallet/UnstoppableWallet/Core/Managers/WalletConnectV2SessionManager.swift
@@ -7,7 +7,7 @@ class WalletConnectV2SessionManager {
     private let disposeBag = DisposeBag()
 
     let service: WalletConnectV2Service
-    private let storage: IWalletConnectV2SessionStorage
+    private let storage: WalletConnectV2SessionStorage
     private let accountManager: IAccountManager
     private let currentDateProvider: ICurrentDateProvider
 
@@ -15,7 +15,7 @@ class WalletConnectV2SessionManager {
     private let pendingRequestsRelay = BehaviorRelay<[Request]>(value: [])
     private let sessionRequestReceivedRelay = PublishRelay<WalletConnectRequest>()
 
-    init(service: WalletConnectV2Service, storage: IWalletConnectV2SessionStorage, accountManager: IAccountManager, currentDateProvider: ICurrentDateProvider) {
+    init(service: WalletConnectV2Service, storage: WalletConnectV2SessionStorage, accountManager: IAccountManager, currentDateProvider: ICurrentDateProvider) {
         self.service = service
         self.storage = storage
         self.accountManager = accountManager
