+++ UnstoppableWallet/UnstoppableWallet/Core/Storage/StorageMigrator.swift
@@ -7,16 +7,9 @@ import KeychainAccess
 import HsToolKit
 import MarketKit
 
-class GrdbStorage {
-    private let dbPool: DatabasePool
+class StorageMigrator {
 
-    init(dbPool: DatabasePool) {
-        self.dbPool = dbPool
-
-        try! migrator.migrate(dbPool)
-    }
-
-    var migrator: DatabaseMigrator {
+    static func migrate(dbPool: DatabasePool) throws {
         var migrator = DatabaseMigrator()
 
         migrator.registerMigration("createAccountRecordsTable") { db in
@@ -597,375 +590,7 @@ class GrdbStorage {
             }
         }
 
-        return migrator
-    }
-
-}
-
-extension GrdbStorage: IEnabledWalletStorage {
-
-    var enabledWallets: [EnabledWallet] {
-        try! dbPool.read { db in
-            try EnabledWallet.fetchAll(db)
-        }
-    }
-
-    func enabledWallets(accountId: String) -> [EnabledWallet] {
-        try! dbPool.read { db in
-            try EnabledWallet.filter(EnabledWallet.Columns.accountId == accountId).fetchAll(db)
-        }
-    }
-
-    func handle(newEnabledWallets: [EnabledWallet], deletedEnabledWallets: [EnabledWallet]) {
-        _ = try! dbPool.write { db in
-            for enabledWallet in newEnabledWallets {
-                try enabledWallet.insert(db)
-            }
-            for enabledWallet in deletedEnabledWallets {
-                try EnabledWallet.filter(EnabledWallet.Columns.coinId == enabledWallet.coinId && EnabledWallet.Columns.coinSettingsId == enabledWallet.coinSettingsId && EnabledWallet.Columns.accountId == enabledWallet.accountId).deleteAll(db)
-            }
-        }
-
-    }
-
-    func clearEnabledWallets() {
-        _ = try! dbPool.write { db in
-            try EnabledWallet.deleteAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: IAccountRecordStorage {
-
-    var allAccountRecords: [AccountRecord] {
-        return try! dbPool.read { db in
-            try AccountRecord.fetchAll(db)
-        }
-    }
-
-    func save(accountRecord: AccountRecord) {
-        _ = try! dbPool.write { db in
-            try accountRecord.insert(db)
-        }
-    }
-
-    func deleteAccountRecord(by id: String) {
-        _ = try! dbPool.write { db in
-            try AccountRecord.filter(AccountRecord.Columns.id == id).deleteAll(db)
-        }
-    }
-
-    func deleteAllAccountRecords() {
-        _ = try! dbPool.write { db in
-            try AccountRecord.deleteAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: IAppVersionRecordStorage {
-
-    var appVersionRecords: [AppVersionRecord] {
-        try! dbPool.read { db in
-            try AppVersionRecord.fetchAll(db)
-        }
-    }
-
-    func save(appVersionRecords: [AppVersionRecord]) {
-        _ = try! dbPool.write { db in
-            for record in appVersionRecords {
-                try record.insert(db)
-            }
-        }
-    }
-
-}
-
-extension GrdbStorage: IFavoriteCoinRecordStorage {
-
-    var favoriteCoinRecords: [FavoriteCoinRecord] {
-        try! dbPool.read { db in
-            try FavoriteCoinRecord.fetchAll(db)
-        }
-    }
-
-    func save(favoriteCoinRecord: FavoriteCoinRecord) {
-        _ = try! dbPool.write { db in
-            try favoriteCoinRecord.insert(db)
-        }
-    }
-
-    func save(favoriteCoinRecords: [FavoriteCoinRecord]) {
-        _ = try! dbPool.write { db in
-            for record in favoriteCoinRecords {
-                try record.insert(db)
-            }
-        }
-    }
-
-    func deleteFavoriteCoinRecord(coinUid: String) {
-        _ = try! dbPool.write { db in
-            try FavoriteCoinRecord
-                    .filter(FavoriteCoinRecord.Columns.coinUid == coinUid)
-                    .deleteAll(db)
-        }
-    }
-
-    func favoriteCoinRecordExists(coinUid: String) -> Bool {
-        try! dbPool.read { db in
-            try FavoriteCoinRecord
-                    .filter(FavoriteCoinRecord.Columns.coinUid == coinUid)
-                    .fetchCount(db) > 0
-        }
-    }
-
-    var favoriteCoinRecords_v_0_22: [FavoriteCoinRecord_v_0_22] {
-        try! dbPool.read { db in
-            try FavoriteCoinRecord_v_0_22.fetchAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: ILogRecordStorage {
-
-    func logs(context: String) -> [LogRecord] {
-        try! dbPool.read { db in
-            try LogRecord
-                    .filter(LogRecord.Columns.context.like("\(context)%"))
-                    .order(LogRecord.Columns.date.asc)
-                    .fetchAll(db)
-        }
-    }
-
-    func save(logRecord: LogRecord) {
-        _ = try? dbPool.write { db in
-            try logRecord.insert(db)
-        }
-    }
-
-    func logsCount() -> Int {
-        try! dbPool.read { db in
-            try LogRecord.fetchCount(db)
-        }
-    }
-
-    func removeFirstLogs(count: Int) {
-        _ = try! dbPool.write { db in
-            let logs = try LogRecord.order(LogRecord.Columns.date.asc).limit(count).fetchAll(db)
-            if let last = logs.last {
-                try LogRecord.filter(LogRecord.Columns.date <= last.date).deleteAll(db)
-            }
-        }
-    }
-
-}
-
-extension GrdbStorage: IWalletConnectSessionStorage {
-
-    func sessions(accountId: String, chainIds: [Int]) -> [WalletConnectSession] {
-        try! dbPool.read { db in
-            try WalletConnectSession.filter(WalletConnectSession.Columns.accountId == accountId && chainIds.contains(WalletConnectSession.Columns.chainId)).fetchAll(db)
-        }
-    }
-
-    func session(peerId: String, accountId: String) -> WalletConnectSession? {
-        try! dbPool.read { db in
-            try WalletConnectSession.filter(WalletConnectSession.Columns.peerId == peerId && WalletConnectSession.Columns.accountId == accountId).fetchOne(db)
-        }
-    }
-
-    func save(session: WalletConnectSession) {
-        _ = try! dbPool.write { db in
-            try session.insert(db)
-        }
-    }
-
-    func deleteSession(peerId: String) {
-        _ = try! dbPool.write { db in
-            try WalletConnectSession.filter(WalletConnectSession.Columns.peerId == peerId).deleteAll(db)
-        }
-    }
-
-    func deleteSessions(accountId: String) {
-        _ = try! dbPool.write { db in
-            try WalletConnectSession.filter(WalletConnectSession.Columns.accountId == accountId).deleteAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: IWalletConnectV2SessionStorage {
-
-    func sessionsV2(accountId: String?) -> [WalletConnectV2Session] {
-        try! dbPool.read { db in
-            var request = WalletConnectV2Session.all()
-            if let accountId = accountId {
-                request = request.filter(WalletConnectV2Session.Columns.accountId == accountId)
-            }
-            return try request.fetchAll(db)
-        }
-    }
-
-    func save(sessions: [WalletConnectV2Session]) {
-        _ = try! dbPool.write { db in
-            for session in sessions {
-                try session.insert(db)
-            }
-        }
-    }
-
-    func deleteSessionV2(topics: [String]) {
-        _ = try! dbPool.write { db in
-            for topic in topics {
-                try WalletConnectV2Session.filter(WalletConnectV2Session.Columns.topic == topic).deleteAll(db)
-            }
-        }
-    }
-
-    func deleteSessionsV2(accountId: String) {
-        _ = try! dbPool.write { db in
-            try WalletConnectV2Session.filter(WalletConnectV2Session.Columns.accountId == accountId).deleteAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: IActiveAccountStorage {
-
-    var activeAccountId: String? {
-        get {
-            try! dbPool.read { db in
-                try ActiveAccount.fetchOne(db)?.accountId
-            }
-        }
-        set {
-            _ = try! dbPool.write { db in
-                if let accountId = newValue {
-                    try ActiveAccount(accountId: accountId).insert(db)
-                } else {
-                    try ActiveAccount.deleteAll(db)
-                }
-            }
-        }
-    }
-
-}
-
-extension GrdbStorage: IRestoreSettingsStorage {
-
-    func restoreSettings(accountId: String, coinId: String) -> [RestoreSettingRecord] {
-        try! dbPool.read { db in
-            try RestoreSettingRecord.filter(RestoreSettingRecord.Columns.accountId == accountId && RestoreSettingRecord.Columns.coinId == coinId).fetchAll(db)
-        }
-    }
-
-    func restoreSettings(accountId: String) -> [RestoreSettingRecord] {
-        try! dbPool.read { db in
-            try RestoreSettingRecord.filter(RestoreSettingRecord.Columns.accountId == accountId).fetchAll(db)
-        }
-    }
-
-    func save(restoreSettingRecords: [RestoreSettingRecord]) {
-        _ = try! dbPool.write { db in
-            for record in restoreSettingRecords {
-                try record.insert(db)
-            }
-        }
-    }
-
-    func deleteAllRestoreSettings(accountId: String) {
-        _ = try! dbPool.write { db in
-            try RestoreSettingRecord.filter(RestoreSettingRecord.Columns.accountId == accountId).deleteAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: IEnabledWalletCacheStorage {
-
-    func enabledWalletCaches(accountId: String) -> [EnabledWalletCache] {
-        try! dbPool.read { db in
-            try EnabledWalletCache.filter(EnabledWalletCache.Columns.accountId == accountId).fetchAll(db)
-        }
-
-    }
-
-    func save(enabledWalletCaches: [EnabledWalletCache]) {
-        _ = try! dbPool.write { db in
-            for cache in enabledWalletCaches {
-                try cache.insert(db)
-            }
-        }
-    }
-
-    func deleteEnabledWalletCaches(accountId: String) {
-        _ = try! dbPool.write { db in
-            try EnabledWalletCache.filter(EnabledWalletCache.Columns.accountId == accountId).deleteAll(db)
-        }
-    }
-
-}
-
-extension GrdbStorage: ICustomTokenStorage {
-
-    func customTokens(platformType: PlatformType, filter: String) -> [CustomToken] {
-        try! dbPool.read { db in
-            try CustomToken
-                    .filter(CustomToken.Columns.coinName.like("%\(filter)%") || CustomToken.Columns.coinCode.like("%\(filter)%"))
-                    .filter(platformType.coinTypeIdPrefixes.map { CustomToken.Columns.coinTypeId.like("\($0)%") }.joined(operator: .or))
-                    .fetchAll(db)
-        }
-    }
-
-    func customTokens(filter: String) -> [CustomToken] {
-        try! dbPool.read { db in
-            try CustomToken
-                    .filter(CustomToken.Columns.coinName.like("%\(filter)%") || CustomToken.Columns.coinCode.like("%\(filter)%"))
-                    .order(CustomToken.Columns.coinName.asc)
-                    .fetchAll(db)
-        }
-    }
-
-    func customTokens(coinTypeIds: [String]) -> [CustomToken] {
-        try! dbPool.read { db in
-            try CustomToken
-                    .filter(coinTypeIds.contains(CustomToken.Columns.coinTypeId))
-                    .fetchAll(db)
-        }
-    }
-
-    func customToken(coinType: MarketKit.CoinType) -> CustomToken? {
-        try! dbPool.read { db in
-            try CustomToken
-                    .filter(CustomToken.Columns.coinTypeId == coinType.id)
-                    .fetchOne(db)
-        }
-    }
-
-    func save(customTokens: [CustomToken]) {
-        _ = try! dbPool.write { db in
-            for customToken in customTokens {
-                try customToken.insert(db)
-            }
-        }
-    }
-
-}
-
-extension GrdbStorage: IEvmAccountSyncStateStorage {
-
-    func evmAccountSyncState(accountId: String, chainId: Int) -> EvmAccountSyncState? {
-        try! dbPool.read { db in
-            try EvmAccountSyncState.filter(EvmAccountSyncState.Columns.accountId == accountId && EvmAccountSyncState.Columns.chainId == chainId).fetchOne(db)
-        }
-    }
-
-    func save(evmAccountSyncState: EvmAccountSyncState) {
-        _ = try! dbPool.write { db in
-            try evmAccountSyncState.insert(db)
-        }
+        try migrator.migrate(dbPool)
     }
 
 }
