+++ UnstoppableWallet/UnstoppableWallet/Core/Storage/WalletConnectSessionStorage.swift
@@ -1,0 +1,44 @@
+import GRDB
+
+class WalletConnectSessionStorage {
+    private let dbPool: DatabasePool
+
+    init(dbPool: DatabasePool) {
+        self.dbPool = dbPool
+    }
+
+}
+
+extension WalletConnectSessionStorage {
+
+    func sessions(accountId: String, chainIds: [Int]) -> [WalletConnectSession] {
+        try! dbPool.read { db in
+            try WalletConnectSession.filter(WalletConnectSession.Columns.accountId == accountId && chainIds.contains(WalletConnectSession.Columns.chainId)).fetchAll(db)
+        }
+    }
+
+    func session(peerId: String, accountId: String) -> WalletConnectSession? {
+        try! dbPool.read { db in
+            try WalletConnectSession.filter(WalletConnectSession.Columns.peerId == peerId && WalletConnectSession.Columns.accountId == accountId).fetchOne(db)
+        }
+    }
+
+    func save(session: WalletConnectSession) {
+        _ = try! dbPool.write { db in
+            try session.insert(db)
+        }
+    }
+
+    func deleteSession(peerId: String) {
+        _ = try! dbPool.write { db in
+            try WalletConnectSession.filter(WalletConnectSession.Columns.peerId == peerId).deleteAll(db)
+        }
+    }
+
+    func deleteSessions(accountId: String) {
+        _ = try! dbPool.write { db in
+            try WalletConnectSession.filter(WalletConnectSession.Columns.accountId == accountId).deleteAll(db)
+        }
+    }
+
+}
