+++ UnstoppableWallet/UnstoppableWallet/Core/Workers/RestoreCustomTokenWorker.swift
@@ -8,12 +8,12 @@ class RestoreCustomTokenWorker {
 
     private let coinManager: CoinManager
     private let walletManager: WalletManager
-    private let storage: IEnabledWalletStorage
+    private let storage: EnabledWalletStorage
     private let localStorage: StorageKit.ILocalStorage
     private let networkManager: NetworkManager
     private let disposeBag = DisposeBag()
 
-    init(coinManager: CoinManager, walletManager: WalletManager, storage: IEnabledWalletStorage, localStorage: StorageKit.ILocalStorage, networkManager: NetworkManager) {
+    init(coinManager: CoinManager, walletManager: WalletManager, storage: EnabledWalletStorage, localStorage: StorageKit.ILocalStorage, networkManager: NetworkManager) {
         self.coinManager = coinManager
         self.walletManager = walletManager
         self.storage = storage
