+++ UnstoppableWallet/UnstoppableWallet/Modules/WalletConnect/WalletConnectSessionManager.swift
@@ -3,13 +3,13 @@ import RxSwift
 import RxRelay
 
 class WalletConnectSessionManager {
-    private let storage: IWalletConnectSessionStorage
+    private let storage: WalletConnectSessionStorage
     private let accountManager: IAccountManager
     private let disposeBag = DisposeBag()
 
     private let sessionsRelay = BehaviorRelay<[WalletConnectSession]>(value: [])
 
-    init(storage: IWalletConnectSessionStorage, accountManager: IAccountManager) {
+    init(storage: WalletConnectSessionStorage, accountManager: IAccountManager) {
         self.storage = storage
         self.accountManager = accountManager
